More refactoring TransparentWnd.h, FixedRegion.h and RecorderView.h

+++ Recorder/RecorderView.h
@@ -220,7 +220,7 @@ private:
     CVideoWnd m_vanWnd;
     CCamera m_cCamera;
     CBasicMessage *m_basicMsg;
-    double _zoom;
+    long long _zoom;
     CPoint _zoomedAt;
     DWORD _zoomWhen;
     int _zoomDirection;
